requirements.txt
streamlit
tensorflow
fastapi
uvicorn
numpy
opencv-python-headless
pillow
python-multipart


gunicorn
sqlalchemy
psycopg2-binary
